# docs/swagger.yml
swagger: '2.0'
info:
    title: User API
    description: Happy to code User API
    version: '1.0.0'
schemes:
    - http
basePath: /api/v1
produces:
    - application/json
paths:
    # /member-home-info/mad/create:
    #     post:
    #         summary: Add a new MAD Member's Confesstion
    #         description: |
    #             Add a new MAD Member's Confesstion
    #         parameters:
    #             - name: data
    #               in: body
    #               description: json body data to add a new member's Confesstion
    #               required: true
    #               schema:
    #                   required:
    #                       - name
    #                       - short_description
    #                       - profile_src
    #                   properties:
    #                       name:
    #                           type: string
    #                           description: Name
    #                       short_description:
    #                           type: string
    #                           description: Short_description of member
    #                       profile_src:
    #                           type: string
    #                           description: Profile_src of member
    #         responses:
    #             200:
    #                 description: Returns a access token
    #             401:
    #                 description: Unauthorized
    #             400:
    #                 description: Bad Request
    #             406:
    #                 description: Not Acceptable
    #             500:
    #                 description: Server Internal error
    #         tags:
    #             - MemberHomeInfoMAD
    /member-home-info/mad/list/{page}/{limit}:
        get:
            summary: Get list of member info
            description: |
                Get list of MAD's member info with page and limit
            parameters:
                - name: page
                  in: path
                  description: Page Number
                  required: true
                  type: string
                  format: string
                - name: limit
                  in: path
                  description: Limit Number
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - MemberHomeInfoMAD
    /member-home-info/mad/get/{id}:
        get:
            summary: Get a MAD's Member
            description: |
                Get a MAD's Member Id
            parameters:
                - name: id
                  in: path
                  description: MAD's Member Id
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - MemberHomeInfoMAD
    # /member-home-info/mad/update/{id}:
    #         put:
    #             summary: Update a MAD's Member
    #             description: |
    #                 Update a MAd's Member with method put
    #             parameters:
    #                 - name: id
    #                   in: path
    #                   description: User Id
    #                   required: true
    #                   type: string
    #                   format: string
    #                 - name: data
    #                   in: body
    #                   description: json body data to update user
    #                   required: true
    #                   schema:
    #                       required:
    #                           - short_description
    #                       properties:
    #                           name:
    #                               type: string
    #                               description: Name
    #                           short_description:
    #                               type: string
    #                               description: Short description of Member
    #                           profile_src:
    #                               type: string
    #                               description: Profile_src of Member
    #             responses:
    #                 200:
    #                     description: OK
    #                 401:
    #                     description: Unauthorized
    #                 400:
    #                     description: Bad Request
    #                 406:
    #                     description: Not Acceptable
    #                 500:
    #                     description: Server Internal error
    #             tags:
    #               - MemberHomeInfoMAD
    /member-home-info/mad/{id}:
        delete:
            summary: Delete a MAD's Member
            description: |
                Delete a MAd's Member with method put
            parameters:
                - name: id
                  in: path
                  description: User Id
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
              - MemberHomeInfoMAD
    # /member-home-info/c4e/create:
    #     post:
    #         summary: Add a new C4E Member's Confesstion
    #         description: |
    #             Add a new C4E Member's Confesstion
    #         parameters:
    #             - name: data
    #               in: body
    #               description: json body data to add a new member's Confesstion
    #               required: true
    #               schema:
    #                   required:
    #                       - name
    #                       - short_description
    #                       - profile_src
    #                   properties:
    #                       name:
    #                           type: string
    #                           description: Name
    #                       short_description:
    #                           type: string
    #                           description: Short_description of member
    #                       profile_src:
    #                           type: string
    #                           description: Profile_src of member
    #         responses:
    #             200:
    #                 description: Returns a access token
    #             401:
    #                 description: Unauthorized
    #             400:
    #                 description: Bad Request
    #             406:
    #                 description: Not Acceptable
    #             500:
    #                 description: Server Internal error
    #         tags:
    #             - MemberHomeInfoC4E
    /member-home-info/c4e/list/{page}/{limit}:
        get:
            summary: Get list of member info
            description: |
                Get list of C4E's member info with page and limit
            parameters:
                - name: page
                  in: path
                  description: Page Number
                  required: true
                  type: string
                  format: string
                - name: limit
                  in: path
                  description: Limit Number
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - MemberHomeInfoC4E
    /member-home-info/c4e/get/{id}:
        get:
            summary: Get a C4E's Member
            description: |
                Get a C4E's Member Id
            parameters:
                - name: id
                  in: path
                  description: MAD's Member Id
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - MemberHomeInfoC4E
    # /member-home-info/c4e/update/{id}:
    #         put:
    #             summary: Update a C4E's Member
    #             description: |
    #                 Update a C4E's Member with method put
    #             parameters:
    #                 - name: id
    #                   in: path
    #                   description: User Id
    #                   required: true
    #                   type: string
    #                   format: string
    #                 - name: data
    #                   in: body
    #                   description: json body data to update user
    #                   required: true
    #                   schema:
    #                       required:
    #                           - short_description
    #                       properties:
    #                           name:
    #                               type: string
    #                               description: Name
    #                           short_description:
    #                               type: string
    #                               description: Short description of Member
    #                           profile_src:
    #                               type: string
    #                               description: Profile_src of Member
    #             responses:
    #                 200:
    #                     description: OK
    #                 401:
    #                     description: Unauthorized
    #                 400:
    #                     description: Bad Request
    #                 406:
    #                     description: Not Acceptable
    #                 500:
    #                     description: Server Internal error
    #             tags:
    #               - MemberHomeInfoC4E
    /member-home-info/c4e/{id}:
        delete:
            summary: Delete a C4E's Member
            description: |
                Delete a C4E's Member with method put
            parameters:
                - name: id
                  in: path
                  description: User Id
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
              - MemberHomeInfoC4E
    # /member-home-info/web/create:
    #     post:
    #         summary: Add a new Web Member's Confesstion
    #         description: |
    #             Add a new Web Member's Confesstion
    #         parameters:
    #             - name: data
    #               in: body
    #               description: json body data to add a new member's Confesstion
    #               required: true
    #               schema:
    #                   required:
    #                       - name
    #                       - short_description
    #                       - profile_src
    #                   properties:
    #                       name:
    #                           type: string
    #                           description: Name
    #                       short_description:
    #                           type: string
    #                           description: Short_description of member
    #                       profile_src:
    #                           type: string
    #                           description: Profile_src of member
    #         responses:
    #             200:
    #                 description: Returns a access token
    #             401:
    #                 description: Unauthorized
    #             400:
    #                 description: Bad Request
    #             406:
    #                 description: Not Acceptable
    #             500:
    #                 description: Server Internal error
    #         tags:
    #             - MemberHomeInfoWeb
    /member-home-info/web/list/{page}/{limit}:
        get:
            summary: Get list of member info
            description: |
                Get list of Web's member info with page and limit
            parameters:
                - name: page
                  in: path
                  description: Page Number
                  required: true
                  type: string
                  format: string
                - name: limit
                  in: path
                  description: Limit Number
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - MemberHomeInfoWeb
    /member-home-info/web/get/{id}:
        get:
            summary: Get a Web's Member
            description: |
                Get a Web's Member Id
            parameters:
                - name: id
                  in: path
                  description: Web's Member Id
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - MemberHomeInfoWeb
    # /member-home-info/web/update/{id}:
    #         put:
    #             summary: Update a Web's Member
    #             description: |
    #                 Update a C4E's Member with method put
    #             parameters:
    #                 - name: id
    #                   in: path
    #                   description: User Id
    #                   required: true
    #                   type: string
    #                   format: string
    #                 - name: data
    #                   in: body
    #                   description: json body data to update user
    #                   required: true
    #                   schema:
    #                       required:
    #                           - short_description
    #                       properties:
    #                           name:
    #                               type: string
    #                               description: Name
    #                           short_description:
    #                               type: string
    #                               description: Short description of Member
    #                           profile_src:
    #                               type: string
    #                               description: Profile_src of Member
    #             responses:
    #                 200:
    #                     description: OK
    #                 401:
    #                     description: Unauthorized
    #                 400:
    #                     description: Bad Request
    #                 406:
    #                     description: Not Acceptable
    #                 500:
    #                     description: Server Internal error
    #             tags:
    #               - MemberHomeInfoWeb
    /member-home-info/web/{id}:
        delete:
            summary: Delete a Web's Member
            description: |
                Delete a C4E's Member with method put
            parameters:
                - name: id
                  in: path
                  description: User Id
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
              - MemberHomeInfoWeb